See Description for changes
This has been a major change in how the BoardHandler works.
Previously the board handler handled multiplexing multiple boards together. While this allowed for greater flexibility in displaying multiple boards such as the player's board and the player's attack board, it did not fit in with the requriements.

As such this update to the BoardHandler is a major overhaul of how the code functions. It now only handles a single board.

Firstly, the BoardHandler has been renamed to BoardPrinterWrapper to reflec it's new function.

Overview of the functions:

The constructor has been reworked to contain fewer parameters. Originally it contained multiple parameters to have maximum flexibility. Instead most parameters has been set to 1.

The addBoard function has been set to a private function. This means that all functions using the addBoard function needs to be reworked. The addBoard function now updates the output board.

The getCopyBoard function now fetches the latest Board by using the addBoard function.

A new function, print has been added.This simplifies the output board process. By putting true in the parameters it prints the attack board. By putting false in the parameters it prints the board normally.