Added ship class
This class adds the function to lay down ships on a board. It also checks for out of bounds and collisions of other ships when setting it down. It has a is sunk function that checks when the ship is sunk.

When using the class please do it in the following order:

Call the constructor.

Use the setMarker function: this sets the ship_marker which is what denotes a ship and a hit_marker for the is sunk function.

Use the setShip function: this lays down the ship on the board. This function has out of bounds and collisions that will return false when the ship has not been laid down.

Lastly, use the is Sunk function: use this function to check the state of the ship. This function is more efficient than reading through the entire array because it knows the position of the ship and does not need to read the board when the ship has already sunk.