Feature/executive logic (#3)
* --added board file

The board file functions are:

--allows creation of 2D char array with placeholder characters

--add and get char from array

--get dimensions of array

--pass by ref of array

--pass by val of array

* --changes

Fixed comment structure.

* --changes

Java 8 Compatibility

* --changes

Fixed typo.

* --changes

Fixed typo

* --changes

changed getCopy() to getCopyMap()

added getCopyBoard()

* --changes

Made code more coherent

* --changes

Fixed bug where 2D array index flipped

* --added BoardHandler

BoardHandler allows the multiplexing
of multiple boards together.

For instance, it allows to multiplex the player's board and the enemy's hidden board to allow for the player to see both his own board and the enemy's board that only shows hit and misses.

It contains the following functions:

--BoardHandler
This is the constructor, it allows you to set the parameters of the main board such
as its size, it's x and y offset and the index offset

--void create
This is the create function. It allows you to set the size of the boards that it will be taking in and the deafault marker that will show where there are no characters.

--void addBoard
This is the adding function. It allows you to add a single board into the main board. It allows you to select if there is indexing or not. Note that the board is mainly passed by value and that the single board does not affect the main board.

--add element
Although it should not be used often, it allows you to add a single character into the arbitrary space of the board. Note that it does not check for Out of index problems.

--getElement
This allows you to retrieve a single character from the main board. Note that it does not check for out of index problems.

--getCopyBoard
This returns a copy of the board. Note that this is pass by value.

* --See Desc. for more info

This quick fix update adds functionality to BoardHandler. It adds the following files:

BoardInterface.java
BoardInterface.class

It also updates the existing code:

BoardHandler.java

with comments.

That's it.

* Added CollisionHandler

This simple class handles detection of collisions inside the board with arbitrary characters and out of bounds detection.

Examples of when to use CollisionHandler:

When inputing a ship onto a board, call CollisionHandler on each spot the ship might occupy before altering the board. CollisionHandler will return true if it detects either a collision with another ship (that has the same character) and it reaches out of bounds of the board.

That's it!

* Se desc. for more info

The changes made to the following classes involved making sure that in order to recieve data from class, the person utilizing the class must invoke the get functions. Previously, you can just grab the data directly.

* Update executive.java

-add commented code outline

* Squashed some comments.

Fixed some of the wording of the comments.

Major thanks to Sean for pointing out that there is no pass by reference and that the behavor is by passing in the address space of the Object.

* Squashed some comments

Use tsopeter/feature/BoardHandler for the most up-to-date version of the Board.java file

* Squashed some comments

Use tsopeter/feature/BoardHandler for the most up-to-date version of Board.java

* Worked on testing user input/output

* Updated gitignore

* Fixed gitignore

* Delete Battleship.class

We don't need to track .class files in the repository

* Delete Board.class

We don't need to track .class files in the repository

* Delete BoardHandler.class

We don't need to track .class files in the repository

* Delete BoardInterface.class

We don't need to track .class files in the repository

* Delete CollisionHandler.class

We don't need to track .class files in the repository

* Delete executive.class

We don't need to track .class files in the repository

* updated helloworld

* Added JavaDoc comment example

Updated Documentation to include an example of proper formatting for a JavaDoc method comment

* Updated clearTerminal() documentation

* See Description for changes

This has been a major change in how the BoardHandler works.
Previously the board handler handled multiplexing multiple boards together. While this allowed for greater flexibility in displaying multiple boards such as the player's board and the player's attack board, it did not fit in with the requriements.

As such this update to the BoardHandler is a major overhaul of how the code functions. It now only handles a single board.

Firstly, the BoardHandler has been renamed to BoardPrinterWrapper to reflec it's new function.

Overview of the functions:

The constructor has been reworked to contain fewer parameters. Originally it contained multiple parameters to have maximum flexibility. Instead most parameters has been set to 1.

The addBoard function has been set to a private function. This means that all functions using the addBoard function needs to be reworked. The addBoard function now updates the output board.

The getCopyBoard function now fetches the latest Board by using the addBoard function.

A new function, print has been added.This simplifies the output board process. By putting true in the parameters it prints the attack board. By putting false in the parameters it prints the board normally.

* Fixed bugs

* Added new feature

This is a rebranded BoardHandler. Use this instead of BoardHandler.

From boardhandler commit:

This has been a major change in how the BoardHandler works.
Previously the board handler handled multiplexing multiple boards together. While this allowed for greater flexibility in displaying multiple boards such as the player's board and the player's attack board, it did not fit in with the requriements.

As such this update to the BoardHandler is a major overhaul of how the code functions. It now only handles a single board.

Firstly, the BoardHandler has been renamed to BoardPrinterWrapper to reflec it's new function.

Overview of the functions:

The constructor has been reworked to contain fewer parameters. Originally it contained multiple parameters to have maximum flexibility. Instead most parameters has been set to 1.

The addBoard function has been set to a private function. This means that all functions using the addBoard function needs to be reworked. The addBoard function now updates the output board.

The getCopyBoard function now fetches the latest Board by using the addBoard function.

A new function, print has been added.This simplifies the output board process. By putting true in the parameters it prints the attack board. By putting false in the parameters it prints the board normally.

* Add files via upload

* Update shot.java

add the Board function

* Update shot.java

* Moved the code to the code folder

* Deleting class files that were pulled in from another branch

* Removed static from method signatures

* Deleting shot.java from my branch because it cannot compile

* Preliminary work on the main game logic

* Update executive.java

* Commented out an import

I will be using it later

* Created and implimented methods to safely handle user input

* More error handling and validating user input

* Implemented the clearTerminal() method

* Created method placeShip()

* Major update to game functionality

* Fixed check()

* Added functionality to the menu options.

* Created and Implemented playerWon()

* Important ship placement fix

Co-authored-by: Tso, Peter <tsopeter@ku.edu>
Co-authored-by: Sam Gilchrist <samofgil@ku.edu>
Co-authored-by: TaoYang666 <70820944+TaoYang666@users.noreply.github.com>
Co-authored-by: samofgil <70820268+samofgil@users.noreply.github.com>